+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -5,17 +5,20 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.booking.Booking.Status;
 import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.BookingDtoFull;
 import ru.practicum.shareit.booking.mapper.BookingMapper;
 import ru.practicum.shareit.booking.service.BookingService;
+import ru.practicum.shareit.booking.service.BookingServiceImpl;
 
 import javax.validation.Valid;
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+import static ru.practicum.shareit.booking.Booking.Status.APPROVED;
+import static ru.practicum.shareit.booking.Booking.Status.REJECTED;
 
 @RestController
 @RequiredArgsConstructor
@@ -40,8 +43,8 @@ public class BookingController {
         BookingDto bookingDto = new BookingDto();
         bookingDto.setId(bookingId);
         bookingDto.setBookerId(bookerId);
-        if (approved) bookingDto.setStatus(Status.APPROVED);
-        else bookingDto.setStatus(Status.REJECTED);
+        if (approved) bookingDto.setStatus(APPROVED);
+        else bookingDto.setStatus(REJECTED);
         return bookingMapper.toDto(bookingService.updateBooking((bookingMapper.toBooking(bookingDto))));
     }
 
@@ -52,7 +55,7 @@ public class BookingController {
 
     @GetMapping
     public List<BookingDtoFull> getBookingInfo(
-            @RequestParam(required = false, defaultValue = "ALL") State state,
+            @RequestParam(required = false, defaultValue = "ALL") BookingServiceImpl.State state,
             @RequestHeader("X-Sharer-User-Id") int bookerId) {
         return Optional.ofNullable(bookingService.getBookings(bookerId, state, false))
                 .stream().flatMap(Collection::stream).map(bookingMapper::toDto).collect(Collectors.toList());
@@ -60,13 +63,9 @@ public class BookingController {
 
     @GetMapping("/owner")
     public List<BookingDtoFull> getBookingInfoOfItems(
-            @RequestParam(required = false, defaultValue = "ALL") State state,
+            @RequestParam(required = false, defaultValue = "ALL") BookingServiceImpl.State state,
             @RequestHeader("X-Sharer-User-Id") int ownerId) {
         return Optional.ofNullable(bookingService.getBookings(ownerId, state, true))
                 .stream().flatMap(Collection::stream).map(bookingMapper::toDto).collect(Collectors.toList());
-    }
-
-    public enum State {
-        ALL, CURRENT, PAST, FUTURE, WAITING, REJECTED
     }
 }
