+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -1,6 +1,5 @@
 package ru.practicum.shareit.item.service;
 
-import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Comment;
 import ru.practicum.shareit.item.model.Item;
 
@@ -12,8 +11,6 @@ public interface ItemService {
     Item updateItem(Item item);
 
     Item getItemById(Integer id);
-
-    ItemDto getItemById(Integer id, Integer userId);
 
     List<Item> getItems(int ownerId);
